Task1-Server/requirements.txt
asgiref==3.8.1
django==5.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pyjwt==1.7.1
sqlparse==0.5.1
